+++ js/main.js
@@ -1,11 +1,9 @@
-import {createKekstaPosts} from './data.js';
-import {NUMBER_OF_KEKSTAPOSTS} from './settings.js';
+import {createKekstaPosts, NUMBER_OF_KEKSTAPOSTS} from './data.js';
+import { renderKekstaPost } from './render.js';
 import './editor.js';
 import './forms.js';
 import './photo_filter.js';
 import './uploader.js';
-
-createKekstaPosts(NUMBER_OF_KEKSTAPOSTS);
-//const generatedKekstaPostsArray = createKekstaPosts(NUMBER_OF_KEKSTAPOSTS);
-
+const generatedKekstaPostsArray = createKekstaPosts(NUMBER_OF_KEKSTAPOSTS);
+renderKekstaPost(generatedKekstaPostsArray);
